Moved remaining session token logic to its module
